+++ opentelemetry-sdk/src/trace/provider.rs
@@ -167,8 +167,8 @@ impl TracerProvider {
     }
 
     /// Create a new [`TracerProvider`] builder.
-    pub fn builder() -> Builder {
-        Builder::default()
+    pub fn builder() -> TracerProviderBuilder {
+        TracerProviderBuilder::default()
     }
 
     /// Span processors associated with this provider
@@ -274,12 +274,12 @@ impl opentelemetry::trace::TracerProvider for TracerProvider {
 
 /// Builder for provider attributes.
 #[derive(Debug, Default)]
-pub struct Builder {
+pub struct TracerProviderBuilder {
     processors: Vec<Box<dyn SpanProcessor>>,
     config: crate::trace::Config,
 }
 
-impl Builder {
+impl TracerProviderBuilder {
     /// Adds a [SimpleSpanProcessor] with the configured exporter to the pipeline.
     ///
     /// # Arguments
@@ -295,7 +295,7 @@ impl Builder {
         let mut processors = self.processors;
         processors.push(Box::new(SimpleSpanProcessor::new(Box::new(exporter))));
 
-        Builder { processors, ..self }
+        TracerProviderBuilder { processors, ..self }
     }
 
     /// Adds a [BatchSpanProcessor] with the configured exporter to the pipeline.
@@ -329,7 +329,7 @@ impl Builder {
         let mut processors = self.processors;
         processors.push(Box::new(processor));
 
-        Builder { processors, ..self }
+        TracerProviderBuilder { processors, ..self }
     }
 
     /// The sdk [`crate::trace::Config`] that this provider will use.
@@ -338,7 +338,7 @@ impl Builder {
         note = "Config is becoming a private type. Use Builder::with_{config_name}(resource) instead. ex: Builder::with_resource(resource)"
     )]
     pub fn with_config(self, config: crate::trace::Config) -> Self {
-        Builder { config, ..self }
+        TracerProviderBuilder { config, ..self }
     }
 
     /// Specify the sampler to be used.
@@ -398,7 +398,7 @@ impl Builder {
     ///
     /// [Tracer]: opentelemetry::trace::Tracer
     pub fn with_resource(self, resource: Resource) -> Self {
-        Builder {
+        TracerProviderBuilder {
             config: self.config.with_resource(resource),
             ..self
         }
@@ -552,7 +552,7 @@ mod tests {
                 provider
                     .config()
                     .resource
-                    .get(Key::from_static_str(resource_key))
+                    .get(&Key::from_static_str(resource_key))
                     .map(|v| v.to_string()),
                 expect.map(|s| s.to_string())
             );
@@ -562,15 +562,18 @@ mod tests {
                 provider
                     .config()
                     .resource
-                    .get(TELEMETRY_SDK_LANGUAGE.into()),
+                    .get(&TELEMETRY_SDK_LANGUAGE.into()),
                 Some(Value::from("rust"))
             );
             assert_eq!(
-                provider.config().resource.get(TELEMETRY_SDK_NAME.into()),
+                provider.config().resource.get(&TELEMETRY_SDK_NAME.into()),
                 Some(Value::from("opentelemetry"))
             );
             assert_eq!(
-                provider.config().resource.get(TELEMETRY_SDK_VERSION.into()),
+                provider
+                    .config()
+                    .resource
+                    .get(&TELEMETRY_SDK_VERSION.into()),
                 Some(Value::from(env!("CARGO_PKG_VERSION")))
             );
         };